# StarkSavingsBank-Backend/src/main/resources/static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Stark Savings Bank API Documentation"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/send-registration": {
            "post": {
                "tags": [
                    "Registration API"
                ],
                "description": "Send the registration email",
                "operationId": "sendRegistrationLink",
                "requestBody": {
                    "description": "Registration Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailResetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/save-account": {
            "post": {
                "tags": [
                    "Registration API"
                ],
                "description": "Save account information",
                "operationId": "saveAccountDetails",
                "requestBody": {
                    "description": "Account Details Request DTO",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountDetailsRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Registration API"
                ],
                "description": "Save the new user",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Customer information Object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EmailResetRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                }
            },
            "AccountDetailsRequestDTO": {
                "type": "object",
                "properties": {
                    "initialDeposit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "branchAddress": {
                        "type": "string"
                    },
                    "accountType": {
                        "type": "string"
                    }
                }
            },
...
            "Customer": {
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 5,
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobileNumber": {
                        "type": "string"
                    },
                    "houseAffiliation": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "CUSTOMER",
                            "EMPLOYEE"
                        ]
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "confirmPassword": {
                        "type": "string",
                        "writeOnly": true
                    }
                }
            },
...
        }
    }
}